+++ pkg/media/pipeline.go
@@ -8,6 +8,7 @@ import (
 	"github.com/tinyzimmer/go-gst/gst"
 
 	"github.com/livekit/ingress/pkg/config"
+	"github.com/livekit/ingress/pkg/params"
 	"github.com/livekit/protocol/livekit"
 	"github.com/livekit/protocol/logger"
 	"github.com/livekit/protocol/tracer"
@@ -15,7 +16,7 @@ import (
 )
 
 type Pipeline struct {
-	*Params
+	*params.Params
 
 	// gstreamer
 	pipeline *gst.Pipeline
@@ -27,7 +28,7 @@ type Pipeline struct {
 	closed         core.Fuse
 }
 
-func New(ctx context.Context, conf *config.Config, params *Params) (*Pipeline, error) {
+func New(ctx context.Context, conf *config.Config, params *params.Params) (*Pipeline, error) {
 	ctx, span := tracer.Start(ctx, "Pipeline.New")
 	defer span.End()
